+++ clare/clare/retry/retry_policy_builder.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from . import stop_strategies
+from . import continue_strategies
 from .retry_policy import RetryPolicy
 
 
@@ -9,11 +9,13 @@ class RetryPolicyBuilder(object):
     def __init__(self,
                  stop_strategies=None,
                  wait_strategy=None,
+                 continue_strategies=None,
                  handled_exceptions=None):
 
         """
         Retry policies should have at least 1 stop strategy and at
-        least 1 wait strategy.
+        least 1 wait strategy. A "successful" attempt is understood as
+        one where an exception was not thrown within the callable.
 
         Parameters
         ----------
@@ -21,18 +23,20 @@ class RetryPolicyBuilder(object):
             Defaults to an empty list.
         wait_strategy : IWaitStrategy
             Defaults to None.
+        continue_strategies : iterable of IContinueStrategy
+            Continuing takes precedence over stopping after successful
+            attempts. In other words, it "overrides" those cases.
+            Defaults to an empty list.
         handled_exceptions : iterable of Exception
             Defaults to an empty tuple.
         """
 
-        self._stop_strategies = stop_strategies
+        self._stop_strategies = stop_strategies or list()
         self._wait_strategy = wait_strategy
-        self._handled_exceptions = handled_exceptions
+        self._continue_strategies = continue_strategies or list()
+        self._handled_exceptions = handled_exceptions or tuple()
 
     def with_stop_strategy(self, stop_strategy):
-        if not self._stop_strategies:
-            self._stop_strategies = list()
-            self._stop_strategies.append(stop_strategies.AfterSuccess())
         self._stop_strategies.append(stop_strategy)
         return self
 
@@ -40,14 +44,21 @@ class RetryPolicyBuilder(object):
         self._wait_strategy = wait_strategy
         return self
 
+    def _with_continue_strategy(self, continue_strategy):
+        self._continue_strategies.append(continue_strategy)
+        return self
+
     def continue_on_exception(self, exception):
-        if not self._handled_exceptions:
-            self._handled_exceptions = tuple()
         self._handled_exceptions += (exception,)
         return self
+
+    def continue_if_result(self, predicate):
+        continue_strategy = continue_strategies.AfterResult(predicate=predicate)
+        return self._with_continue_strategy(continue_strategy)
 
     def build(self):
         retry_policy = RetryPolicy(stop_strategies=self._stop_strategies,
                                    wait_strategy=self._wait_strategy,
+                                   continue_strategies=self._continue_strategies,
                                    handled_exceptions=self._handled_exceptions)
         return retry_policy