+++ clare/documentation/release_notes.md
@@ -10,10 +10,10 @@
 - Added stopping after attempt (`ME-659`).
 - Added stopping after duration (`ME-659`).
 - Added stopping after never (`ME-659`).
-- Added stopping after result (`ME-659`).
 - Added stopping on multiple conditions (`ME-659`).
 - Added waiting for fixed duration (`ME-659`).
 - Added continuing on exceptions (`ME-659`).
+- Added continuing on results (`ME-659`).
 
 ## v0.2.0 (Pending)
 ##### Changes